+++ pkg/gen/input/apptest/internal/file/values.go
@@ -1,0 +1,23 @@
+package file
+
+import (
+	_ "embed"
+	"path/filepath"
+
+	"github.com/giantswarm/devctl/v6/pkg/gen/input"
+	"github.com/giantswarm/devctl/v6/pkg/gen/input/apptest/internal/params"
+)
+
+//go:embed values.yaml.template
+var createValuesTemplate string
+
+func NewCreateValuesInput(p params.Params) input.Input {
+	i := input.Input{
+		Path:           filepath.Join(p.Dir, "suites/basic", "values.yaml"),
+		TemplateBody:   createValuesTemplate,
+		TemplateData:   map[string]interface{}{},
+		SkipRegenCheck: true,
+	}
+
+	return i
+}